+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Reranking based on text + abstract + title.
+- Match on text + title.
 
 ## [0.0.11] - 05.05.2025